ITDS-25 refactor: 소셜 로그인 Factory Pattern 적용
