+++ src/main/java/com/dissonance/itit/service/OAuthService.java
@@ -1,0 +1,10 @@
+package com.dissonance.itit.service;
+
+import com.dissonance.itit.domain.enums.SocialLoginProvider;
+import com.dissonance.itit.dto.response.OAuthUserInformation;
+
+public interface OAuthService {
+	SocialLoginProvider getProvider();
+
+	OAuthUserInformation requestUserInformation(String token);
+}
